Add case sensitive username
Kerberos is case sensitive, Windows is not. Before this change `require`
statements were returned with the sAMAccountName in lowercase. That is
preferred if users are typing the username themselves but when using
Windows and SSO Kerberos negotiation of browsers the username is sent
with the casing as entered in Active Directory.

This change returns both the lowercase username and the actual account
name as is from Active Directory if they are not identical.